requirements.txt
slackclient
requests
imageio
opencv-python
scapy
backoff
pygifsicle
